requirements.txt
streamlit
groq
python-dotenv
streamlit-audiorec
streamlit-audiorecorder